Update totalSalience for keywords in swiped articles to be used for sorting

+++ app/src/main/java/com/csc306/coursework/model/AnalyzeEntitiesResponse.kt
@@ -6,5 +6,5 @@ class AnalyzeEntitiesResponse(
 
 class Entity(
     val name: String,
-    val salience: Float
+    val salience: Double
 )
